Try to isolate controller logic.
App will have a bunch of controllers. Each controllers wil be called in a responsability
chain. Until a controller confirms it manages the current request, the app will invoke a
next controller.
The controller::do_routing is the method a controller must implement. The implementation
should return true if the request is successfully recognised.

Use the account MVC package in the Blog app.
Fix namespaces to belong MVC packages logic.