Mustache: move render functions in a processor
Define an interface for Mustache escapers, and provide an adapter for HTML render escaper
and a null escaper.
Change the parser to use state objects for tags, instead of constants.
Adapt tests to support relaxed helpers (that takes anything that can cast in a h\string)
and to use the new processor helper.