Lib: _thow becaume _exception
Encapsulate throw in a single place is not a good idea, as it makes backtraces illigible.
Besides, the actual site from the exception is raised is lost, debugging becomming a pain.
I although rename unit test method _excpetion_thrown and _exception_not_thrown to avoid
confusion with exception factory methods.